llm_api/pyproject.toml
[tool.poetry]
name = "llm_api"
version = "0.1.0"
description = ""
authors = [

]

maintainers = [

]

readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = { version = "^0.22.0", extras = ["standard"] }
pydantic = "^2"
pydantic-settings = "^2"
openai = "^1.61.1"
ujson = "^5.10.0"

[tool.poetry.dev-dependencies]
pytest = "^8.2"
psycopg2-binary = "2.9.10"
pytest-asyncio = "^0.25.3"
pytest-mock = "3.14.0"
flake8 = "~7.0.0"
black = "^24.4.2"
autoflake = "^1.6.1"
pyyaml = "^6.0.2"
requests = "^2.32.3"

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"